Fyp CFMS/backend/requirements.txt
Django== 5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
PyMySQL==1.1.0
cryptography==41.0.7
python-decouple==3.8
drf-yasg==1.21.7
Pillow==11.0.0
PyPDF2==3.0.1
pandas==2.2.0
openpyxl==3.1.2
